[swappable] Simplified tests
